LayoutTests/platform/ios/imported/w3c/web-platform-tests/XMLHttpRequest/send-entity-body-none-expected.txt


PASS No content type (POST) 
PASS No content type (PUT) 
FAIL No content type (HEAD) assert_equals: expected "NO" but got "0"
PASS Explicit content type (POST) 
PASS Explicit content type (PUT) 
FAIL Explicit content type (HEAD) assert_equals: expected "NO" but got "0"
